Move formatting tests in
