+++ gov_uk_dashboards/formatting/number_formatting.py
